Add RaySpawner and Film. Cast dummy rays for each pixel

+++ src/render/frame_data.rs
@@ -1,10 +1,10 @@
 use crate::math::Color;
 
 pub struct FrameData {
-    pub w: i32,
-    pub h: i32,
+    pub w: usize,
+    pub h: usize,
     pub pixels: Vec<Color>,
-    pub frame_number: i32,
+    pub frame_number: usize,
     pub t: f64,
     // TODO: camera settings (for debugging)
 }
